+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to IG Publisher Manager will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.9.0] - 2025-29-08
+### Added
+- Text Filter
+- Drop IGs from Finder/Explorer etc 
+
+### Changed
+- Check for missing folders
+
 ## [v0.8.0] - 2025-22-08
 ### Added
 - Search by IG